Finish basic processing
Though, some categorical features still remain that need to be converted to numerical and some converted ones need to be one hot encoded also we need to clean the file and not forget to 'to use statistical tools that you’ve learned, such as statistical tests, to find out complex relationships in your dataset'. Furthermore, we may be able to gather more data for example by scraping unavailable digikala phones that are available in emalls.